Split data and datalist class into modules
