+++ src/modules/data.js
@@ -1,0 +1,6 @@
+export default class Data {
+  constructor(name, score) {
+    this.name = name;
+    this.score = score;
+  }
+}
